Fix some issues outlined in the pull request: * function_to_fdatabasis to _utils * docstring in each std function * use pytest instead of unittest
