+++ skfda/_utils/_utils.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import functools
 import numbers
-from functools import singledispatch
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -36,7 +35,7 @@ from ._sklearn_adapter import BaseEstimator
 ArrayDTypeT = TypeVar("ArrayDTypeT", bound="np.generic")
 
 if TYPE_CHECKING:
-    from ..representation import FData, FDataGrid
+    from ..representation import FData, FDataBasis, FDataGrid
     from ..representation.basis import Basis
     from ..representation.extrapolation import ExtrapolationLike
 
@@ -609,3 +608,36 @@ def _classifier_get_classes(
             f'one; got {classes.size} class',
         )
     return classes, y_ind
+
+
+def function_to_fdatabasis(
+    f: Callable[[NDArrayFloat], NDArrayFloat],
+    new_basis: Basis,
+) -> FDataBasis:
+    """Express a math function as a FDataBasis with a given basis.
+
+    Args:
+        f: math function.
+        new_basis: the basis of the output.
+
+    Returns:
+        FDataBasis: FDataBasis with calculated coefficients and the new
+        basis.
+    """
+    from .. import FDataBasis
+    from ..misc._math import inner_product_matrix
+
+    if isinstance(f, FDataBasis) and f.basis == new_basis:
+        return f
+
+    inner_prod = inner_product_matrix(
+        new_basis,
+        f,
+        _domain_range=new_basis.domain_range,
+    )
+
+    gram_matrix = new_basis.gram_matrix()
+
+    coefs = np.linalg.solve(gram_matrix, inner_prod)
+
+    return FDataBasis(new_basis, coefs.T)
